+++ ke-legacy/main.js
@@ -27,6 +27,8 @@ var clientEnrollment = require('../client-enrollment/clientEnrollment');
 var justInTime = require('../just-in-time/justInTime');
 var rosterAPI = require('../rw-legacy/lib/roster/api');
 var slacLogger = require('../slack-logger/index');
+var dukaClient = require('../duka-client/dukaClient');
+var isCreditOfficer = require('../duka-client/checkCreditOfficer');
 if(service.active){
     defaultEnvironment = 'prod';
 }else{
@@ -47,6 +49,8 @@ service.vars.lr_2021_client_table_id = project.vars[env+'_lr_2021_client_table_i
 service.vars.registerEnrollEnd = env+ '_registerEnrollEnd';
 service.vars.registerEnrollStart = env+ '_registerEnrollStart';
 var checkGroupLeader = require('../shared/rosterApi/checkForGroupLeader');
+service.vars.credit_officers_table = 'credit_officers_table';
+service.vars.duka_clients_table = env + '_duka_client_registration';
 
 if(env == 'prod'){
     service.vars.topUpBundleTableId = 'DT891c89e9a82b6841';
@@ -1562,6 +1566,7 @@ transactionHistory.registerHandlers();
 clientRegistration.registerHandlers();
 justInTime.registerHandlers();
 groupRepaymentsModule.registerGroupRepaymentHandlers({lang: GetLang() ? 'en' : 'sw', main_menu: state.vars.main_menu, main_menu_handler: 'MainMenu'});
+dukaClient.registerInputHandlers(GetLang() ? 'en-ke' : 'sw', service.vars.duka_clients_table);
 
 
 function reduceClientSize(client) {
@@ -1588,6 +1593,7 @@ addInputHandler('SplashMenu', function(SplashMenu) {
         promptDigits('StaffPayRoll', {submitOnHash: true, maxDigits: 5, timeout: 5});
     }
     else {
+        var creditOfficerDetails = isCreditOfficer(ClientAccNum, service.vars.credit_officers_table);
         if (RosterClientVal(ClientAccNum)){
             console.log('SuccessFully Validated against Roster');
             client = RosterClientGet(ClientAccNum);
@@ -1606,6 +1612,8 @@ addInputHandler('SplashMenu', function(SplashMenu) {
             state.vars.account_number = client.AccountNumber;
             MainMenuText(client);
             promptDigits('MainMenu', {submitOnHash: true, maxDigits: 8, timeout: 5});
+        } if(creditOfficerDetails) {
+            dukaClient.start(GetLang() ? 'en-ke' : 'sw', creditOfficerDetails);
         }
         else{
             console.log('account number not valid');